refactor(app_runtime): wizard issue-level open helpers を wizard.rs に追加 (SPEC-2077 Phase F2a-2)
- crates/gwt/src/app_runtime/wizard.rs に open_issue_launch_wizard_events / handle_issue_launch_wizard_prepared を追加
- mod.rs から該当 method を削除 (~134 行)
- 残: F2b (handle_launch_wizard_action ~600 行) / F3a/F3b (spawn ~525 行)

Verification:
- cargo test -p gwt-core -p gwt: 全 pass (既知 flake 1 件、 個別 pass 確認)
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
